refactor: Consolidate all _utils unit tests of bigframes function
